pyproject.toml
[tool.poetry]
name = "stakewolle"
version = "0.1.0"
description = "Backend test  task from Stakewolle"
authors = ["SergeiNaum <sergei.naumencko@yandex.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2"
djangorestframework = "^3.14.0"
django-extensions = "^3.2.1"
python-dotenv = "^1.0.0"
gunicorn = "^21.2.0"
drf-spectacular = "^0.27.1"
requests = "^2.31.0"
celery = "^5.3.6"
redis = "^5.0.3"
djangorestframework-simplejwt = "^5.3.1"
aiohttp = "^3.9.3"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
black = "^24.1.1"
ipython = "^8.5.0"
flower = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
